Remove 'Solving method' from the solver dialog.
The --method parameter affects only the current "soft thread" (= single
scan), may confuse laymen, and should not matter when an aggregate scan
preset or theme is specified, and those are in general superior and more
encouraged for fc-solve beginners and most experts a like. An aggregate
scan explicitly sets the --method for each scan individually, so there.

As a result, I don't see the point of having the 'Solving method' there.
Perhaps in the future, a more comprehensive GUI for editing the fc-solve
command line will be in order, but that may end up something like thess
monstrocities:

* http://www.codinghorror.com/blog/2006/11/this-is-what-happens-when-you-let-developers-create-ui.html

* http://i.stack.imgur.com/cVWub.gif