Udate cardset_viewer.py
Original doesn't work for me so I redesigned it.
The tool starts in Appdir and searches for valid card sets.
Since not all the sets are always there, I gave him a file list to rummage through the directories. You can also call up the info and the config file. To view the sets, simply click on the name with the mouse or scroll up and down with the arrow keys.